Make the compiler a little bit happier

+++ libc/lib/string/memset.c
@@ -1,10 +1,11 @@
 #include <string.h>
+#include <stdint.h>
 
 void *memset(void *_data, int val, size_t len) {
     // TODO: There is a vast room for optimizations.
     unsigned char *data = _data;
     for(;len; len--) {
-        data[len] = val;
+        data[len] = (uint8_t)val;
     }
     return data;
 }
